pyproject.toml
[tool.poetry]
name = "py-macropad"
version = "0.1.0"
description = ""
authors = ["Yassir Karroum <ukarroum17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
pyalsaaudio = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
